+++ src/main/java/com/space/server/domin/inventory/domain/repository/InventoryRepository.java
@@ -1,6 +1,6 @@
-package com.space.server.core.inventory.domain.repository;
+package com.space.server.domin.inventory.domain.repository;
 
-import com.space.server.core.inventory.domain.Inventory;
+import com.space.server.domin.inventory.domain.Inventory;
 import com.space.server.core.item.domain.Item;
 import com.space.server.core.item.domain.value.Category;
 import com.space.server.user.domain.Users;
