+++ src/main/java/com/space/server/domin/inventory/presentation/dto/response/InventoryResponse.java
@@ -1,6 +1,6 @@
-package com.space.server.core.inventory.presentation.dto.response;
+package com.space.server.domin.inventory.presentation.dto.response;
 
-import com.space.server.core.inventory.domain.Inventory;
+import com.space.server.domin.inventory.domain.Inventory;
 import com.space.server.core.item.domain.value.Category;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
