+++ src/main/java/com/space/server/domin/inventory/service/QueryInventoryService.java
@@ -1,7 +1,7 @@
-package com.space.server.core.inventory.service;
+package com.space.server.domin.inventory.service;
 
-import com.space.server.core.inventory.domain.Inventory;
-import com.space.server.core.inventory.service.implementation.InventoryReader;
+import com.space.server.domin.inventory.domain.Inventory;
+import com.space.server.domin.inventory.service.implementation.InventoryReader;
 import com.space.server.user.domain.Users;
 import com.space.server.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
