+++ src/main/java/com/space/server/domin/inventory/service/implementation/InventoryDeleter.java
@@ -1,7 +1,7 @@
-package com.space.server.core.inventory.service.implementation;
+package com.space.server.domin.inventory.service.implementation;
 
-import com.space.server.core.inventory.domain.Inventory;
-import com.space.server.core.inventory.domain.repository.InventoryRepository;
+import com.space.server.domin.inventory.domain.Inventory;
+import com.space.server.domin.inventory.domain.repository.InventoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
