+++ src/main/java/com/space/server/domin/inventory/service/implementation/InventoryReader.java
@@ -1,8 +1,8 @@
-package com.space.server.core.inventory.service.implementation;
+package com.space.server.domin.inventory.service.implementation;
 
-import com.space.server.core.inventory.domain.Inventory;
-import com.space.server.core.inventory.domain.repository.InventoryRepository;
-import com.space.server.core.inventory.exception.InventoryNotFoundException;
+import com.space.server.domin.inventory.domain.Inventory;
+import com.space.server.domin.inventory.domain.repository.InventoryRepository;
+import com.space.server.domin.inventory.exception.InventoryNotFoundException;
 import com.space.server.core.item.domain.value.Category;
 import com.space.server.user.domain.Users;
 import lombok.RequiredArgsConstructor;
